requirements.txt
pandas
matplotlib
sqlalchemy
argparse
biopython
configparser
